refactor: Clean up Calculator component
Moved a lot of mathematical (non-react) functionality into new
file (logic.js). This is accessed via the evaluateRPN function,
the only function actually needed in Calculator. Also slightly
tweaked some parameters so that they're independent of the
component, e.g. evaluateRPN now takes calcState and answer as
params rather than using their bare values.
